// src/db/shared/schema.ts
import { z } from 'zod';

export type SubType = 'character' | 'datapack' | 'system' | 'theme' | 'subscription'

export type Item = {
  // garanteed unique id, generated by server.
  id: string

  local_id: string // deviceid-uuid
  owner_id: string

  // set via function that takes doc / snapshot and returns object to shadow.
  shadow: { [key:string]: any }

  namespace: string
  type: SubType

  lifecycle: 'crdt' | 'snapshot'

  // CRDT state (only if lifecycle === "draft")
  doc: any | null

  // Immutable artifact state (only if lifecycle === "published")
  snapshot: any | null

  // incremented on server sync, 0 for newly created items
  version: number

  // used solely on the server for change tracking.
  // never to be changed on the client, and will be ignored.
  updated_at: number
  deleted_at: number
  last_change_index: bigint
}

export const itemSchema = z.object({
  id: z.string({ error: 'id must be a string' }).max(100, 'id cannot be longer than 100 characters').optional(),

  local_id: z.string({ error: 'local_id must be a string' }).max(73, 'local_id must be 73 characters or less'),
  user_id: z.string(),

  name: z.string(),

  forked_from: z.string().nullable().optional(),

  // a random hash, a quick glance to see if update conflicts are happening.
  version: z.string().optional(),

  created_at: z.iso.datetime({ offset: true }),
  updated_at: z.iso.datetime({ offset: true }),
})